Completed SQLite Handler & UDP Broadcast finder.
Completed the class for SQLite database handling. Added a way to find all the devices that supports this function via a udp broadcast packet.

Also reorganized a bit the code to make it cleaner.

Still need to do some code cleanup / improvements and need to add an auto-purge of the database based on a configurable max size.